+++ plutus-tx-plugin/test/AsData/Budget/Spec.hs
@@ -1,16 +1,5 @@
-{-# LANGUAGE BangPatterns          #-}
-{-# LANGUAGE DataKinds             #-}
-{-# LANGUAGE FlexibleContexts      #-}
-{-# LANGUAGE LambdaCase            #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE NegativeLiterals      #-}
-{-# LANGUAGE OverloadedStrings     #-}
-{-# LANGUAGE PatternSynonyms       #-}
-{-# LANGUAGE ScopedTypeVariables   #-}
-{-# LANGUAGE TemplateHaskell       #-}
-{-# LANGUAGE TypeApplications      #-}
-{-# LANGUAGE ViewPatterns          #-}
-{-# OPTIONS_GHC -fplugin PlutusTx.Plugin #-}
+{-# LANGUAGE DataKinds       #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:context-level=0 #-}
 {-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:defer-errors #-}
 
@@ -24,7 +13,7 @@ import PlutusTx.Builtins qualified as PlutusTx
 import PlutusTx.Code
 import PlutusTx.IsData qualified as PlutusTx
 import PlutusTx.Lift (liftCodeDef)
-import PlutusTx.Test (goldenBudget, goldenEvalCekCatch, goldenPirReadable)
+import PlutusTx.Test (goldenBudget, goldenEvalCekCatch, goldenPirReadable, goldenUPlcReadable)
 import PlutusTx.TH (compile)
 
 import AsData.Budget.Types
@@ -34,14 +23,21 @@ tests =
   testNestedGhc
     ("AsData" </> "Budget")
     [ goldenPirReadable "onlyUseFirstField" onlyUseFirstField
+    , goldenUPlcReadable "onlyUseFirstField" onlyUseFirstField
     , goldenEvalCekCatch "onlyUseFirstField" $ [onlyUseFirstField `unsafeApplyCode` inp]
     , goldenBudget "onlyUseFirstField-budget" (onlyUseFirstField `unsafeApplyCode` inp)
     , goldenPirReadable "patternMatching" patternMatching
+    , goldenUPlcReadable "patternMatching" patternMatching
     , goldenEvalCekCatch "patternMatching" $ [patternMatching `unsafeApplyCode` inp]
     , goldenBudget "patternMatching-budget" (patternMatching `unsafeApplyCode` inp)
     , goldenPirReadable "recordFields" recordFields
+    , goldenUPlcReadable "recordFields" recordFields
     , goldenEvalCekCatch "recordFields" $ [recordFields `unsafeApplyCode` inp]
     , goldenBudget "recordFields-budget" (recordFields `unsafeApplyCode` inp)
+    , goldenPirReadable "recordFields-manual" recordFieldsManual
+    , goldenUPlcReadable "recordFields-manual" recordFieldsManual
+    , goldenEvalCekCatch "recordFields-manual" $ [recordFieldsManual `unsafeApplyCode` inp]
+    , goldenBudget "recordFields-budget-manual" (recordFieldsManual `unsafeApplyCode` inp)
     ]
 
 -- A function that only accesses the first field of `Ints`.
@@ -60,12 +56,25 @@ patternMatching =
         [||
         \d ->
           let (Ints x y z w) = PlutusTx.unsafeFromBuiltinData d
-           in x `PlutusTx.addInteger` y `PlutusTx.addInteger` z `PlutusTx.addInteger` w
+           in x
+                `PlutusTx.addInteger` y
+                `PlutusTx.addInteger` z
+                `PlutusTx.addInteger` w
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (y `PlutusTx.addInteger` z)
+                                          (x `PlutusTx.addInteger` w)
+                                          then x `PlutusTx.addInteger` z
+                                          else y `PlutusTx.addInteger` w
+                                      )
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (z `PlutusTx.addInteger` y)
+                                          (w `PlutusTx.addInteger` x)
+                                          then z `PlutusTx.addInteger` x
+                                          else w `PlutusTx.addInteger` y
+                                      )
         ||]
     )
 
--- TODO: this does the same thing as `patternMatching`, but is much more expensive,
--- since there's no sharing between the code that accesses different fields.
 recordFields :: CompiledCode (PlutusTx.BuiltinData -> Integer)
 recordFields =
   $$( compile
@@ -76,7 +85,61 @@ recordFields =
               y = int2 ints
               z = int3 ints
               w = int4 ints
-           in x `PlutusTx.addInteger` y `PlutusTx.addInteger` z `PlutusTx.addInteger` w
+           in x
+                `PlutusTx.addInteger` y
+                `PlutusTx.addInteger` z
+                `PlutusTx.addInteger` w
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (y `PlutusTx.addInteger` z)
+                                          (x `PlutusTx.addInteger` w)
+                                          then x `PlutusTx.addInteger` z
+                                          else y `PlutusTx.addInteger` w
+                                      )
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (int3 ints `PlutusTx.addInteger` int2 ints)
+                                          (int4 ints `PlutusTx.addInteger` int1 ints)
+                                          then
+                                            int3 ints
+                                              `PlutusTx.addInteger` int1 ints
+                                          else
+                                            int4 ints
+                                              `PlutusTx.addInteger` int2 ints
+                                      )
+        ||]
+    )
+
+-- This is much more efficient than `recordFields` since the manually written
+-- field accessors are more CSE-friendly.
+recordFieldsManual :: CompiledCode (PlutusTx.BuiltinData -> Integer)
+recordFieldsManual =
+  $$( compile
+        [||
+        \d ->
+          let ints = PlutusTx.unsafeFromBuiltinData d
+              x = int1Manual ints
+              y = int2Manual ints
+              z = int3Manual ints
+              w = int4Manual ints
+           in x
+                `PlutusTx.addInteger` y
+                `PlutusTx.addInteger` z
+                `PlutusTx.addInteger` w
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (y `PlutusTx.addInteger` z)
+                                          (x `PlutusTx.addInteger` w)
+                                          then x `PlutusTx.addInteger` z
+                                          else y `PlutusTx.addInteger` w
+                                      )
+                `PlutusTx.addInteger` ( if PlutusTx.lessThanInteger
+                                          (int3Manual ints `PlutusTx.addInteger` int2Manual ints)
+                                          (int4Manual ints `PlutusTx.addInteger` int1Manual ints)
+                                          then
+                                            int3Manual ints
+                                              `PlutusTx.addInteger` int1Manual ints
+                                          else
+                                            int4Manual ints
+                                              `PlutusTx.addInteger` int2Manual ints
+                                      )
         ||]
     )