+++ plutus-tx/src/PlutusTx/Builtins.hs
@@ -69,6 +69,7 @@ module PlutusTx.Builtins (
                          , pairToPair
                          -- * Lists
                          , matchList
+                         , headMaybe
                          , BI.head
                          , BI.tail
                          , uncons
@@ -84,8 +85,8 @@ module PlutusTx.Builtins (
                          , bls12_381_G1_compress
                          , bls12_381_G1_uncompress
                          , bls12_381_G1_hashToGroup
-                         , bls12_381_G1_zero
-                         , bls12_381_G1_generator
+                         , bls12_381_G1_compressed_zero
+                         , bls12_381_G1_compressed_generator
                          , BuiltinBLS12_381_G2_Element
                          , bls12_381_G2_equals
                          , bls12_381_G2_add
@@ -94,8 +95,8 @@ module PlutusTx.Builtins (
                          , bls12_381_G2_compress
                          , bls12_381_G2_uncompress
                          , bls12_381_G2_hashToGroup
-                         , bls12_381_G2_zero
-                         , bls12_381_G2_generator
+                         , bls12_381_G2_compressed_zero
+                         , bls12_381_G2_compressed_generator
                          , BuiltinBLS12_381_MlResult
                          , bls12_381_millerLoop
                          , bls12_381_mulMlResult
@@ -103,6 +104,8 @@ module PlutusTx.Builtins (
                          -- * Conversions
                          , fromBuiltin
                          , toBuiltin
+                         , builtinIntegerToByteString
+                         , builtinByteStringToInteger
                          ) where
 
 import Data.Maybe
@@ -381,6 +384,10 @@ encodeUtf8 = BI.encodeUtf8
 matchList :: forall a r . BI.BuiltinList a -> r -> (a -> BI.BuiltinList a -> r) -> r
 matchList l nilCase consCase = BI.matchList l (\_ -> nilCase) (\x xs _ -> consCase x xs) ()
 
+{-# INLINE headMaybe #-}
+headMaybe :: BI.BuiltinList a -> Maybe a
+headMaybe l = matchList l Nothing (\h _ -> Just h)
+
 {-# INLINE uncons #-}
 -- | Uncons a builtin list, failing if the list is empty, useful in patterns.
 uncons :: BI.BuiltinList a -> Maybe (a, BI.BuiltinList a)
@@ -510,13 +517,13 @@ bls12_381_G1_uncompress = BI.bls12_381_G1_uncompress
 bls12_381_G1_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G1_Element
 bls12_381_G1_hashToGroup = BI.bls12_381_G1_hashToGroup
 
-{-# INLINABLE bls12_381_G1_zero #-}
-bls12_381_G1_zero :: BuiltinBLS12_381_G1_Element
-bls12_381_G1_zero = BI.bls12_381_G1_zero
+{-# INLINABLE bls12_381_G1_compressed_zero #-}
+bls12_381_G1_compressed_zero :: BuiltinByteString
+bls12_381_G1_compressed_zero = BI.bls12_381_G1_compressed_zero
 
-{-# INLINABLE bls12_381_G1_generator #-}
-bls12_381_G1_generator :: BuiltinBLS12_381_G1_Element
-bls12_381_G1_generator = BI.bls12_381_G1_generator
+{-# INLINABLE bls12_381_G1_compressed_generator #-}
+bls12_381_G1_compressed_generator :: BuiltinByteString
+bls12_381_G1_compressed_generator = BI.bls12_381_G1_compressed_generator
 
 -- G2 --
 {-# INLINABLE bls12_381_G2_equals #-}
@@ -547,13 +554,13 @@ bls12_381_G2_uncompress = BI.bls12_381_G2_uncompress
 bls12_381_G2_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G2_Element
 bls12_381_G2_hashToGroup = BI.bls12_381_G2_hashToGroup
 
-{-# INLINABLE bls12_381_G2_zero #-}
-bls12_381_G2_zero :: BuiltinBLS12_381_G2_Element
-bls12_381_G2_zero = BI.bls12_381_G2_zero
+{-# INLINABLE bls12_381_G2_compressed_zero #-}
+bls12_381_G2_compressed_zero :: BuiltinByteString
+bls12_381_G2_compressed_zero = BI.bls12_381_G2_compressed_zero
 
-{-# INLINABLE bls12_381_G2_generator #-}
-bls12_381_G2_generator :: BuiltinBLS12_381_G2_Element
-bls12_381_G2_generator = BI.bls12_381_G2_generator
+{-# INLINABLE bls12_381_G2_compressed_generator #-}
+bls12_381_G2_compressed_generator :: BuiltinByteString
+bls12_381_G2_compressed_generator = BI.bls12_381_G2_compressed_generator
 
 -- Pairing --
 {-# INLINABLE bls12_381_millerLoop #-}
@@ -567,3 +574,17 @@ bls12_381_mulMlResult = BI.bls12_381_mulMlResult
 {-# INLINABLE bls12_381_finalVerify #-}
 bls12_381_finalVerify :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> Bool
 bls12_381_finalVerify a b = fromBuiltin (BI.bls12_381_finalVerify a b)
+
+-- Conversions
+
+-- | Convert a 'BuiltinInteger' into a 'BuiltinByteString', as described in
+-- [CIP-0087](https://github.com/mlabs-haskell/CIPs/tree/koz/to-from-bytestring/CIP-XXXX).
+{-# INLINEABLE builtinIntegerToByteString #-}
+builtinIntegerToByteString :: Bool -> Integer -> Integer -> BuiltinByteString
+builtinIntegerToByteString endiannessArg = BI.builtinIntegerToByteString (toBuiltin endiannessArg)
+
+-- | Convert a 'BuiltinByteString' to a 'BuiltinInteger', as described in
+-- [CIP-0087](https://github.com/mlabs-haskell/CIPs/tree/koz/to-from-bytestring/CIP-XXXX).
+builtinByteStringToInteger :: Bool -> BuiltinByteString -> Integer
+builtinByteStringToInteger statedEndiannessArg =
+  BI.builtinByteStringToInteger (toBuiltin statedEndiannessArg)